+++ tests/src/FunctionalJavascript/ArticleCreationTest.php
@@ -96,7 +96,7 @@ class ArticleCreationTest extends ThunderJavascriptTestBase {
     $this->getSession()
       ->wait(5000, "jQuery('iframe').filter(function(){return (this.id.indexOf('twitter-widget-0') !== -1);}).length === 1");
     $numOfElements = $this->getSession()->evaluateScript("jQuery('iframe').filter(function(){return (this.id.indexOf('twitter-widget-0') !== -1);}).length");
-    $this->assertEquals(1, $numOfElements, "Number of instagrams on page should be one.");
+    $this->assertEquals(1, $numOfElements, "Number of twitter on page should be one.");
 
     // Check Link Paragraph.
     $this->assertSession()->linkExists('Link to Thunder');
@@ -111,7 +111,7 @@ class ArticleCreationTest extends ThunderJavascriptTestBase {
     $this->getSession()
       ->wait(5000, "jQuery('iframe').filter(function(){return (this.src.indexOf('youtube.com/embed/Ksp5JVFryEg') !== -1);}).length === 1");
     $numOfElements = $this->getSession()->evaluateScript("jQuery('iframe').filter(function(){return (this.src.indexOf('youtube.com/embed/Ksp5JVFryEg') !== -1);}).length");
-    $this->assertEquals(1, $numOfElements, "Number of instagrams on page should be one.");
+    $this->assertEquals(1, $numOfElements, "Number of youtube on page should be one.");
 
     // Check that one Pinterest widget is on page.
     $this->assertSession()
